+++ mitra_adapters/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod accounts;
 pub mod dynamic_config;
 pub mod init;
 mod logger;
@@ -6,4 +7,3 @@ pub mod payments;
 pub mod permissions;
 pub mod posts;
 pub mod profiles;
-pub mod roles;