+++ mitra_cli/src/commands/account.rs
@@ -15,7 +15,8 @@ use clap::{
 
 use mitra_activitypub::adapters::users::create_or_update_local_actor;
 use mitra_adapters::{
-    roles::{
+    accounts::{
+        account_type_to_str,
         from_default_role,
         role_from_str,
         role_to_str,
@@ -25,9 +26,11 @@ use mitra_adapters::{
 use mitra_config::Config;
 use mitra_models::{
     accounts::{
-        helpers::get_user_by_id_or_name,
+        helpers::{
+            create_anonymous_account,
+            get_user_by_id_or_name,
+        },
         queries::{
-            create_automated_account,
             create_invite_code,
             create_user,
             get_accounts_for_admin,
@@ -36,14 +39,11 @@ use mitra_models::{
             set_user_role,
         },
         types::{
-            AutomatedAccountData,
-            AutomatedAccountType,
             UserCreateData,
         },
     },
     database::{get_database_client, DatabaseConnectionPool},
     oauth::queries::delete_oauth_tokens,
-    profiles::types::ANONYMOUS,
 };
 use mitra_utils::passwords::hash_password;
 use mitra_validators::accounts::validate_local_username;
@@ -105,16 +105,11 @@ impl CreateSystemAccount {
     ) -> Result<(), Error> {
         let db_client = &mut **get_database_client(db_pool).await?;
         let instance = config.instance();
-        let account_data = AutomatedAccountData {
-            username: ANONYMOUS.to_owned(),
-            bio: None,
-            bio_source: None,
-            emojis: vec![],
-            account_type: AutomatedAccountType::Anonymous,
-            rsa_secret_key: instance.rsa_secret_key,
-            ed25519_secret_key: instance.ed25519_secret_key,
-        };
-        create_automated_account(db_client, account_data).await?;
+        create_anonymous_account(
+            db_client,
+            instance.rsa_secret_key,
+            instance.ed25519_secret_key,
+        ).await?;
         println!("account created");
         Ok(())
     }
@@ -132,19 +127,16 @@ impl ListAccounts {
         let db_client = &**get_database_client(db_pool).await?;
         let accounts = get_accounts_for_admin(db_client).await?;
         println!(
-            "{0: <40} | {1: <35} | {2: <20} | {3: <35} | {4: <35}",
-            "ID", "username", "role", "created", "last login",
+            "{0: <40} | {1: <35} | {2: <10} | {3: <15} | {4: <35} | {5: <35}",
+            "ID", "username", "type", "role", "created", "last login",
         );
         for account in accounts {
-            let role = match account.role {
-                Some(role) => role_to_str(role),
-                None => "user (portable)",
-            };
             println!(
-                "{0: <40} | {1: <35} | {2: <20} | {3: <35} | {4: <35}",
+                "{0: <40} | {1: <35} | {2: <10} | {3: <15} | {4: <35} | {5: <35}",
                 account.profile.id.to_string(),
                 account.profile.username,
-                role,
+                account_type_to_str(account.account_type),
+                account.role.map(role_to_str).unwrap_or("-"),
                 account.profile.created_at.to_string(),
                 account.last_login.map(|dt| dt.to_string()).unwrap_or_default(),
             );